requirements.txt
asgiref==3.5.2
autoflake==1.6.1
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
django-registration-redux==2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.1
